Implement OlympiadBench evaluation utilities and refactor math verification
- Introduced new utility functions for processing OlympiadBench documents and results in `en_utils.py` and `zh_utils.py`.
- Added a custom timeout decorator in `math_verify_utils.py` to replace the previous signal-based timeout.
- Created multiple YAML configuration files for various OlympiadBench tasks, including math and physics in both English and Chinese.
- Removed outdated files from the `olympiadbench_official` directory to streamline the codebase.
- Enhanced evaluation logic for OlympiadBench tasks in `olympiadbench_evals.py` and added aggregation functions for results processing.
