+++ lmms_eval/tasks/olympiadbench_mimo/olympiadbench_all_boxed.yaml
