+++ lmms_eval/tasks/olympiadbench_mimo/olympiadbench_boxed.yaml
