+++ lmms_eval/tasks/olympiadbench_mimo/olympiadbench_image_math_en.yaml
