+++ lmms_eval/tasks/olympiadbench_mimo/olympiadbench_image_physics_en.yaml
