+++ lmms_eval/tasks/olympiadbench_mimo/olympiadbench_math_en_no_proof.yaml
