+++ lmms_eval/tasks/olympiadbench_mimo/olympiadbench_math_zh.yaml
