+++ lmms_eval/tasks/olympiadbench_mimo/olympiadbench_official.yaml
@@ -1,6 +1,6 @@
 # Copyright 2025 Xiaomi Corporation.
 
-group: olympiadbench_official
+group: olympiadbench_mimo
 task:
 - olympiadbench_math_en
 - olympiadbench_math_zh
