+++ lmms_eval/tasks/olympiadbench_mimo/olympiadbench_official_en_no_proof.yaml
@@ -1,10 +1,8 @@
 # Copyright 2025 Xiaomi Corporation.
 
-group: olympiadbench_official_no_proof
+group: olympiadbench_mimo_en_no_proof
 task:
 - olympiadbench_math_en_no_proof
-- olympiadbench_math_zh_no_proof
 - olympiadbench_physics_en_no_proof
-- olympiadbench_physics_zh_no_proof
 metadata:
   - version: 0.0
