+++ lmms_eval/tasks/olympiadbench_mimo/olympiadbench_official_zh.yaml
@@ -1,6 +1,6 @@
 # Copyright 2025 Xiaomi Corporation.
 
-group: olympiadbench_official_zh
+group: olympiadbench_mimo_zh
 task:
 - olympiadbench_math_zh
 - olympiadbench_physics_zh
