+++ lmms_eval/tasks/olympiadbench_mimo/olympiadbench_official_zh_no_proof.yaml
@@ -1,6 +1,6 @@
 # Copyright 2025 Xiaomi Corporation.
 
-group: olympiadbench_official_zh_no_proof
+group: olympiadbench_mimo_zh_no_proof
 task:
 - olympiadbench_math_zh_no_proof
 - olympiadbench_physics_zh_no_proof
