+++ lmms_eval/tasks/olympiadbench_mimo/olympiadbench_physics_zh.yaml
