+++ lmms_eval/tasks/olympiadbench_mimo/utils.py
@@ -9,7 +9,7 @@ from loguru import logger as eval_logger
 
 from lmms_eval.tasks._task_utils.file_utils import generate_submission_file
 from lmms_eval.tasks._task_utils.math_verify_utils import MathVerifyFn
-from lmms_eval.tasks.olympiadbench_official.olympiadbench_evals import (
+from lmms_eval.tasks.olympiadbench_mimo.olympiadbench_evals import (
     OlympiadBenchEvaluator,
 )
