+++ lmms_eval/tasks/olympiadbench_mimo/zh_utils.py
@@ -7,7 +7,7 @@ import os
 from loguru import logger as eval_logger
 
 from lmms_eval.tasks._task_utils.file_utils import generate_submission_file
-from lmms_eval.tasks.olympiadbench_official.olympiadbench_evals import (
+from lmms_eval.tasks.olympiadbench_mimo.olympiadbench_evals import (
     OlympiadBenchEvaluator,
 )
